requirements.txt
PySide6>=6.4.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6